+++ backend/controllers/slackController.js
@@ -2,7 +2,7 @@ const { slackService } = require('../services/slackService')
 const { slackClient } = require('../services/slackClient')
 const slack = slackService({ slackClient })
 
-const { importHistory, parseTimestamp } = require('../utils/importHistory')
+const { importHistory, parseTimestamp } = require('../application/importHistory')
 
 const getAllByUser = async (res, id) => {
   try {
@@ -33,9 +33,7 @@ const getChannels = async (res) => {
 
 const getChannelHistory = async (res, channel) => {
   try {
-    console.log('DSADASDAS11111 ', channel)
     const result = await importHistory(res, channel)
-    console.log('ch history :  ',result)
     res.json(result)
   } catch (error) {
     res.send(error)
@@ -47,9 +45,7 @@ const getChannelHistoryWithParameters = async (req, res) => {
     var channel = req.body.channel || 'general'
     var oldest = parseTimestamp(Date.now() * 1000, req.body.hours)
     var user = req.body.user
-    console.log('DSADASDAS')
     const result = await importHistory(res, channel, oldest, user)
-    console.log('ch history with param :  ',result)
     res.json(result)
   } catch (error) {
     res.send(error)
